safari-600.7-branch/LayoutTests/fast/selectors/pseudo-class-not-expected.txt

The pseudo class :not

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


:not selector matched.
PASS getComputedStyle(document.getElementById("target1")).backgroundColor is "rgb(0, 0, 0)"
:not selector unmatched.
PASS getComputedStyle(document.getElementById("target2")).backgroundColor is "rgb(1, 2, 3)"
multiple :not selector matched.
PASS getComputedStyle(document.getElementById("target3")).backgroundColor is "rgb(2, 3, 4)"
multiple :not selector matched.
PASS getComputedStyle(document.getElementById("target4")).backgroundColor is "rgb(2, 3, 4)"
multiple :not selector unmatched.
PASS getComputedStyle(document.getElementById("target5")).backgroundColor is "rgb(0, 0, 0)"
:not with backtracking matched.
PASS getComputedStyle(document.getElementById("target6")).backgroundColor is "rgb(1, 2, 3)"
:not with backtracking failed.
PASS getComputedStyle(document.getElementById("target7")).backgroundColor is "rgb(0, 0, 0)"
:not with maximum registers matched.
PASS getComputedStyle(document.getElementById("target8")).backgroundColor is "rgb(1, 2, 3)"
:not with maximum registers matched.
PASS getComputedStyle(document.getElementById("target9")).backgroundColor is "rgb(1, 2, 3)"
PASS successfullyParsed is true

TEST COMPLETE
